+++ hamster_lib/helpers/time.py
@@ -56,7 +56,7 @@ def end_day_to_datetime(end_day, config):
 
     Args:
         end (datetime.date): Raw end date that is to be adjusted.
-        config: Controler config containing information on when a workday starts.
+        config: Controller config containing information on when a workday starts.
 
     Returns:
         datetime.datetime: The endday as a adjusted datetime object.
@@ -287,6 +287,7 @@ def complete_timeframe(timeframe, config, partial=False):
 
     if any((timeframe.end_date, timeframe.end_time)) or not partial:
         end = complete_end(timeframe.end_date, timeframe.end_time, config)
+
     return (start, end)
 
 
@@ -309,10 +310,6 @@ def parse_time(time):
         This parse just a singlular date, time or datetime representation.
     """
 
-    # [TODO]
-    # Propably should enhance error handling for invalid but correctly formated
-    # times such as '30:55' or '2015-15-60'.
-
     length = len(time.strip().split())
     if length == 1:
         try:
@@ -323,6 +320,32 @@ def parse_time(time):
         result = datetime.datetime.strptime(time, '%Y-%m-%d %H:%M')
     else:
         raise ValueError(_(
-            "Sting does not seem to be in one of our supported time formats."
+            "String does not seem to be in one of our supported time formats."
         ))
     return result
+
+
+def validate_start_end_range(range_tuple):
+    """
+    Perform basic sanity checks on a timeframe.
+
+    Args:
+        range_tuple (tuple): ``(start, end)`` tuple as returned by
+            ``complete_timeframe``.
+
+    Raises:
+        ValueError: If validation fails.
+
+    Returns:
+        tuple: ``(start, end)`` tuple that passed validation.
+    """
+
+    start, end = range_tuple
+
+    if not start:
+        raise ValueError(_("Start missing."))
+
+    if (start and end) and (start > end):
+        raise ValueError(_("Start after end!"))
+
+    return range_tuple